Rename categoryConstants to filterConstants
